requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.41
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
Werkzeug==3.1.3